move methods to the PendingRelayDb companion object
